src/requirements.txt
Django==3.0.2
psycopg2==2.8.4
django-filter==2.2.0
djangorestframework==3.11.0
